refactor(broker): extract token persistence to token_store.rs (#588)
Move save_token_map and load_token_map from broker/auth.rs to new
broker/token_store.rs, separating disk I/O from auth logic. Update
imports in admin.rs, mod.rs, and daemon/migration.rs. Move 3 pure
persistence unit tests; keep integration tests in auth.rs.

auth.rs: 843 → 787 lines. token_store.rs: 62 lines.

Closes #588
