+++ crates/room-cli/src/broker/auth.rs
@@ -1,4 +1,3 @@
-use std::collections::HashMap;
 use std::path::Path;
 use std::sync::Arc;
 
@@ -9,31 +8,7 @@ use uuid::Uuid;
 use crate::registry::UserRegistry;
 
 use super::state::{SubscriptionMap, TokenMap};
-
-// ── Token persistence ────────────────────────────────────────────────────────
-
-/// Write a token map to disk as JSON.
-pub(crate) fn save_token_map(map: &HashMap<String, String>, path: &Path) -> Result<(), String> {
-    let json = serde_json::to_string_pretty(&map).map_err(|e| format!("serialize tokens: {e}"))?;
-    std::fs::write(path, json).map_err(|e| format!("write {}: {e}", path.display()))
-}
-
-/// Load a token map from disk. Returns an empty map if the file does not exist.
-///
-/// `token_map_path` is the `.tokens` file (see [`crate::paths::broker_tokens_path`]).
-pub(crate) fn load_token_map(token_map_path: &Path) -> HashMap<String, String> {
-    let contents = match std::fs::read_to_string(token_map_path) {
-        Ok(c) => c,
-        Err(_) => return HashMap::new(),
-    };
-    serde_json::from_str(&contents).unwrap_or_else(|e| {
-        eprintln!(
-            "[auth] corrupt token file {}: {e}",
-            token_map_path.display()
-        );
-        HashMap::new()
-    })
-}
+use super::token_store::save_token_map;
 
 /// Check whether `username` is allowed to join a room with the given config.
 ///
@@ -273,6 +248,7 @@ pub(crate) async fn handle_oneshot_join_with_registry(
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::broker::token_store::load_token_map;
     use room_protocol::{RoomConfig, RoomVisibility};
     use std::{collections::HashMap, sync::Arc};
     use tokio::sync::Mutex;
@@ -506,39 +482,7 @@ mod tests {
         assert!(check_send_permission("anyone", None).is_ok());
     }
 
-    // ── Token persistence ─────────────────────────────────────────────────
-
-    #[test]
-    fn load_token_map_missing_file_returns_empty() {
-        let dir = tempfile::tempdir().unwrap();
-        let token_map_path = dir.path().join("nonexistent.tokens");
-        let map = load_token_map(&token_map_path);
-        assert!(map.is_empty());
-    }
-
-    #[test]
-    fn save_and_load_round_trip() {
-        let dir = tempfile::tempdir().unwrap();
-        let token_map_path = dir.path().join("test.tokens");
-
-        let mut original = HashMap::new();
-        original.insert("tok-1".to_owned(), "alice".to_owned());
-        original.insert("tok-2".to_owned(), "bob".to_owned());
-
-        save_token_map(&original, &token_map_path).unwrap();
-        let loaded = load_token_map(&token_map_path);
-        assert_eq!(loaded, original);
-    }
-
-    #[test]
-    fn load_token_map_corrupt_file_returns_empty() {
-        let dir = tempfile::tempdir().unwrap();
-        let token_map_path = dir.path().join("corrupt.tokens");
-        std::fs::write(&token_map_path, "not json{{{").unwrap();
-
-        let map = load_token_map(&token_map_path);
-        assert!(map.is_empty());
-    }
+    // ── Token persistence integration ───────────────────────────────────
 
     #[tokio::test]
     async fn issue_token_persists_to_disk() {
